Changed blog_id type from long to integer so now works correctly

+++ Project/RefugEAP-Project/src/main/java/com/example/refugeapproject/membership/Blog.java
@@ -9,16 +9,16 @@ public class Blog {
     @Id
     @Column(name = "blog_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long blog_id;
+    private Integer blog_id;
 
     private String title;
     private String content;
 
-    public Long getBlog_id() {
+    public Integer getBlog_id() {
         return blog_id;
     }
 
-    public void setBlog_id(Long blog_id) {
+    public void setBlog_id(Integer blog_id) {
         this.blog_id = blog_id;
     }
